Switch from link-check to html-proofer
The html-proofer gem is still maintained.

To test the build instead of testing against he live site, I had to
rework the travis.yml file.  The sleep statement is to make sure that
the Jekyll server is up and running by the time we try to follow the
URLs that point to other places on the localhost.